+++ filling-core/src/main/java/com/filling/calculation/plugin/base/flink/source/ElasticSearchSource.java
@@ -70,7 +70,7 @@ public class ElasticSearchSource implements FlinkStreamSource<Row> {
     @Override
     public CheckResult checkConfig() {
         if(!elasticsearchConf.getAutoSchema()) {
-            return CheckConfigUtil.check(config, SCHEMA);
+            return CheckConfigUtil.check(config);
         }
         CheckResult result = new CheckResult(true, "");
         return result;
