+++ filling-core/src/main/java/com/filling/calculation/plugin/base/flink/source/elasticsearch/ElasticsearchConf.java
@@ -72,7 +72,7 @@ public class ElasticsearchConf {
     /** table field names */
     private String[] fieldNames;
 
-    private Boolean autoSchema = false;
+    private Boolean autoSchema = true;
 
     public int getMaxConnPerRoute() {
         return maxConnPerRoute;
