+++ filling-core/src/main/java/com/filling/calculation/plugin/base/flink/source/elasticsearch/ElasticsearchUtil.java
@@ -23,7 +23,7 @@ import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
 
-import static org.apache.flink.streaming.connectors.elasticsearch.table.ElasticsearchOptions.HOSTS_OPTION;
+import static org.apache.flink.streaming.connectors.elasticsearch.table.ElasticsearchConnectorOptions.HOSTS_OPTION;
 import static org.elasticsearch.index.query.QueryBuilders.matchAllQuery;
 
 /**
