+++ filling-core/src/main/java/com/filling/calculation/plugin/base/flink/transform/DecodeBase64.java
@@ -10,7 +10,6 @@ import com.filling.calculation.flink.util.TableUtil;
 import org.apache.flink.streaming.api.datastream.DataStream;
 import org.apache.flink.table.api.Table;
 import org.apache.flink.table.api.TableEnvironment;
-import org.apache.flink.table.api.bridge.java.BatchTableEnvironment;
 import org.apache.flink.table.api.bridge.java.StreamTableEnvironment;
 import org.apache.flink.types.Row;
 
@@ -40,7 +39,7 @@ public class DecodeBase64 implements FlinkStreamTransform<Row, Row> {
             .replaceAll("\\{source_field}", config.getString(SOURCE_FIELD_NAME))
             .replaceAll("\\{target_field}", config.getString(TARGET_FIELD_NAME));
         Table table = tableEnvironment.sqlQuery(sql).dropColumns(config.getString(SOURCE_FIELD_NAME));
-        return "batch".equals(type) ? TableUtil.tableToDataSet((BatchTableEnvironment) tableEnvironment, table) : TableUtil.tableToDataStream((StreamTableEnvironment) tableEnvironment, table, false);
+        return TableUtil.tableToDataStream((StreamTableEnvironment) tableEnvironment, table, false);
     }
 
     @Override
