+++ filling-core/src/test/resources/flink/es2es.json
@@ -8,10 +8,10 @@
     {
       "plugin_name": "ElasticSearchSource",
       "result_table_name": "es01",
-      "hosts": ["10.10.14.51:9200"],
-      "index": "mca-68",
+      "hosts": ["192.168.100.203:9201"],
+      "index": "internal",
       "name": "es-source",
-      "schema": "{\"begintime\": \"123\"}"
+      "parallelism": 10
     }
   ],
   "transform": [],
@@ -20,10 +20,10 @@
       "source_table_name": "es01",
       "plugin_name": "Elasticsearch",
       "hosts": [
-        "10.10.14.51:9200"
+        "192.168.1.242:9200"
       ],
-      "index": "test003",
-      "parallelism": 1
+      "index": "test004",
+      "parallelism": 9
     }
   ]
 }
